server: Add structures to describe the player state
The player state is now stored in an array of fv_player objects
maintained by an object called fv_playerbase. The array is stored as
pointers to the players so that the addresses won't change when more
players are added and the connection can keep the pointer to its
player. The players can be indexed internally by the player number
(which is just the position in the array) or a global player ID. The
intention is that the players can be moved around within the array
and the player numbers will be updated accordingly.