common: Add a utility for maintaining an fv_buffer of bits
This adds utility functions for handling an fv_buffer containing a
series of bits. The bits are stored grouped in unsigned longs. Sizing
the buffer ensures that the size is rounded up to the nearest long
with the padding bits unset. There is a loop macro which can quickly
iterate over the bits using ffsl.