client/network: Keep track of the other players' state
The network object now handles the messages from the server to
maintain an array of fv_person objects representing the other players.
A bitmask is kept with a bit for each player that is modified.
Whenever a consistent state is reached a callback is invoked to report
the new status. This is called asynchronously and reports a pointer to
the player state with the bitmask of modified players.