client/logic: Report a more fine-grained state change
Instead of just returning a bool from fv_logic_update, it now returns
a set of flags so that the caller can distinguish between things that
need updating on the network and things that just cause a redraw. That
way it will later be possible for the client to go sleep if nothing is
happening.