server/player-base: Automatically remove stale players
The playerbase now has a timeout to periodically check for players
that don't have a client haven't had any data from a client for more
than 2 minutes. If so they are removed. This can potentially cause the
dirty signal to be emitted because the last player in the array will
be moved to fill in the gap created by the missing player. The dirty
event now includes a bool to say whether the number of players has
changed.