client: Send keep alive messages
In order to prevent the server from thinking the connection is stale,
the client now sends data at least every minute. If no other data
needs to be sent, it will send an empty KEEP_ALIVE message. This is a
new message type which has been added to the protocol document.