server/connection: Write output commands in a WebSocket frame
If the connection is talking via a WebSocket it will now prepend each
command with a WebSocket frame which marks a complete message. To do
this the fv_proto_write_command function is now wrapped with a
function that prepends a two-byte frame header. The payload length
therefore must now always be less than 126.