Use the WebSocket protocol directly instead of trying to filter
The fv_ws_parser is now only used to parse the initial HTTP headers
sent as part of the WebSocket protocol. Once the connection is
established then fv_connection directly handles parsing the frames.
The fv_proto_* utility functions now directly generate WebSocket
messages. The protocol document is changed to remove the length for
the messages because this is now deduced from the length of the
WebSocket message.

This makes the code a bit simpler and it should make it easier to
write the browser side.

fv_connection also now theoretically handles command frames in the
middle of fragmented messages. It should also send pong replies to
ping frames.

Additionally the size of the message opcode field has been changed to
a single byte.