client: Add a minimum reconnect wait time
Otherwise if the server doesn't like the client for some reason and
repeatedly closes the connection immediately then the client will
flood the server with connections. This limits it to at most 1
connection per second. The reconnect time is measured from the last
connect call so if the connection dies after a long successful
connection then it will effectively reconnect immediately, which is
probably the right thing to do.