emscripten/main: Pause the main loop when nothing is happening
When the logic becomes stable it now pauses the main loop instead of
repeatedly painting the same scene. The previous mechanism to queue a
redraw by setting a flag has now been encapsulated in a function so
that the emscripten version can resume the main loop there. Events are
now handled by installing an event filter instead of just before
painting because otherwise an event will not cause the main loop to be
iterated.