server: Handle speech packets
When a client sends a speech packet it is now added to a circular
buffer. There is now a number in the dirty state for each player to
mark how many of the speeches need to be sent to the client. They are
then copied to the client when filling the write buffer. The server
doesn't do any encoding or decoding of speeches, instead they are just
checked for validity and then copied verbatim.