client: Add a microphone object
The microphone object is a simple wrapper around the microphone which
invokes a callback whenever data is received from the mic. The native
version launches a thread and uses the simple PulseAudio API to do
blocking reads for the data. The Emscripten version is currently just
a stub which never reports any data.