client: Add a buffer for encoding audio data
This adds an fv_recorder object which has a circular buffer of encoded
data. It creates an fv_microphone object and listens for data to
encode. The buffer grows to at most 3 seconds of data. When a new
packet is added it invokes a callback. The processing is done in a
separate thread so the callback function needs to be thread safe.
There are accessor functions to check whether a packet is available
and retrieve the next one from the queue.