client: Add an object to buffer and mix uncompressed audio
Adds fv_audio_buffer which can be passed compressed audio packets. It
will then decompress them and mix them with a circular buffer of data.
There is a function to consume data from the buffer. If there is not
enough data in the buffer then silence is filled for the remainder.