client/map-painter: Support specials with a texture
There are now two programs used to render a special, one with only
solid colors and one with a texture and texture coordinates. The
original texture program has been renamed to be a map program. The
specials data now includes a texture image number as well as the name
of the model.