client: Encode the specials at the bottom of the map image
Instead of just having a single pixel to describe the one special at a
particular block, the specials are now encoded with machine-readable
data at the bottom of the image for the map. The assumption is that
now there is an editor there is less need to be able to edit the
specials with an image editor. However the editor still sets the
ignored colours to something so that it is clear where the specials
are. This should make it easier to later have multiple specials per
block.