client: Handle touch events
Touch events are now handled as alternative way to move the player.
This just updates what was previously called the mouse_screen_position
similar to how the mouse is handled. These variables have now been
renamed to cursor_*.

This has only been lightly tested on Firefox with Emscripten. It's not
really clear whether the Emscripten SDL driver correctly reports the
finger positions because according to the documentation they are
supposed to be normalised to 0..1 but this isn't actually happening.