Check for -std=c11 and -std=gnu11 instead of just -std=c99
Otherwise if something includes sys/cdefs.h glibc will provide a macro
to emulate _Static_assert because it thinks it isn't available. The
macro won't work in the middle of a struct definition so the
build breaks. This also changes it to set CFLAGS so that the
subsequent check for _Static_assert will use this flag.