client/logic: Use fv_logic_person as the base for player and npc
Previously the two structs fv_logic_player and fv_logic_person had a
lot of common data which equated to fv_logic_person but it was stored
separately. When the persons are iterated it would then copy this
information to a temporary fv_logic_person struct. This is a bit
wasteful and will get more awkward once more state is added to
fv_logic_person. Instead of doing this both fv_logic_player and
fv_logic_npc now have a person field which directly stores the person
data.

In the old game which this source code was based on there was more
common code to describe a position which was required by both players
and npcs in order to calculate the movement. However, in Babiling only
the players need this data so it was redundant in the npc struct.
Therefore the fv_logic_position struct has now been removed and its
members are just directly stored in the fv_logic_player struct.