client: Don't use the network types for fv_person
Instead of using the stdint types that are passed over the network in
fv_person, they are now stored directly as floats as used by fv_logic.
It is now only the responsibility of fv_network to convert between the
formats, which makes more sense because that is really only a property
of the network protocol and doesn't really concern the rest of the
layers.