client/network: Add an assert that all of the player state gets sent
If we make it past the code to send the player state in fill_write_buf
then we should definitely not have any remaining flags. If there were
any left then it would probably continuously wake up the main loop.
This might happen if some player state is added without adding the
code to send it.