client/logic: Keep track of the person to show flags for
Adds fv_logic_get/set_flag_person to access the person. Instead of
storing the person number the details of the person are copied. This
is because the npc player numbers can change when people enter or
leave the game. Sadly this means the flags won't update live while the
menu is open.